requirements.txt
black
pandas
pymssql
setuptools
pyodbc
azure-cognitiveservices-speech
streamlit
openai
plotly
langchain
langchain_openai
langchain_core
python-dotenv